Update in voice assistant code
Created separate files for voice assistant window, speech to text code and the main function.

The voice assistant window contains the code for buttons to record user question and/or exit program.

The speech to text file contains the code that listens to user response and convert it to text.

The main file runs the whole program.